feat: support mnist test
